Fix Rolling includes (#47)
* Fix Rolling includes


* Update library_generation/library_generation.sh


* Rolling workarround

+++ library_generation/library_generation.sh
@@ -47,10 +47,15 @@ else
 fi
 
 ######## Fix include paths  ########
-INCLUDE_ROS2_PACKAGES=( rmw rcl rcl_action rcl_lifecycle rcl_logging_interface rosidl_runtime_c rosidl_typesupport_interface)
-for var in "${INCLUDE_ROS2_PACKAGES[@]}"; do
-    mv $BASE_PATH/libmicroros/include/${var}/${var}/* $BASE_PATH/libmicroros/include/${var}
-    rm -rf $BASE_PATH/libmicroros/include/${var}/${var}
+pushd $BASE_PATH/libmicroros/micro_ros_src > /dev/null
+    INCLUDE_ROS2_PACKAGES=$(colcon list | awk '{print $1}' | awk -v d=" " '{s=(NR==1?s:s d)$0}END{print s}')
+popd > /dev/null
+
+for var in ${INCLUDE_ROS2_PACKAGES}; do
+    if [ -d "$BASE_PATH/libmicroros/include/${var}/${var}" ]; then
+        rsync -r $BASE_PATH/libmicroros/include/${var}/${var}/* $BASE_PATH/libmicroros/include/${var}
+        rm -rf $BASE_PATH/libmicroros/include/${var}/${var}
+    fi
 done
 
 ######## Generate extra files ########
